+++ resources/views/partials/user-actions.blade.php
@@ -1,11 +1,14 @@
-<div class="flex items-center justify-end gap-1">
-    <x-bt-button 
-        xs 
-        ghost 
-        icon-start="cog-6-tooth"
-        wire:click="$dispatch('manageUserAssignments', { id: {{ $row['id'] }} })"
-        :title="__('beartropy-permissions::messages.manage_roles_and_permissions')"
+<div class="flex items-center justify-end gap-1 mr-1">
+    <x-bt-tooltip
+        :label="__('beartropy-permissions::messages.manage_roles_and_permissions')"
     >
-        {{ __('beartropy-permissions::messages.manage') }}
-    </x-bt-button>
+        <x-bt-button 
+            xs 
+            ghost    
+            amber
+            wire:click="$dispatch('manageUserAssignments', { id: {{ $row['id'] }} })"
+        >
+            <x-bt-icon name="cog-6-tooth" />
+        </x-bt-button>
+    </x-bt-tooltip>
 </div>