+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.1.1] - 2026-01-16
+
+### Fixed
+- Fixed ID column alignment in Users, Roles, and Permissions tables.
+
+### Changed
+- Cleaned up blade views for actions.
+
 ## [v0.1.0] - 2026-01-15
 
 ### Added